+++ test/integration/server.cc
@@ -12,6 +12,7 @@
 #include "source/common/stats/thread_local_store.h"
 #include "source/common/thread_local/thread_local_impl.h"
 #include "source/server/hot_restart_nop_impl.h"
+#include "source/server/instance_impl.h"
 #include "source/server/options_impl.h"
 #include "source/server/process_context_impl.h"
